feat(core): Exports as ESM only, supports browsers
Browsers will now be detected and support for colors will be loaded async and only be printed if the bundler setup supports it.

The styling (colors, thorough pretty-printing) can be configured (turned on/off), although this should be less of an issue with the browser detection.

BREAKING CHANGE: Requires ESM and Node >=14.13.1

fix #13
fix #21
fix #22